fixed bug:
in the media stream provider we've got

if (getResource()->hasUnprocessedCommands()) // if command processor has something in the queue for this resource let it go first

So if resource has UnprocessedCommands we do not request image for it.

from other hand we just got one processor for all resources.

so if one resource is offline and we post some command for another, another will wait till the command for first_offline resource is executed.
while waiting image is not requested.


wait time could take a while coz of network timeout.

1) Network timeout is reduced
2) if commanf failed once, I remove all comand for such resource in a queue.
3) todo: introduce a pool of processors instead of one

Shortly speaking:

if you drop multimple cameras on layout. one of them is online => the others starts show video much faster( This is for arecont only so far )

<==== Mercurial data ====>
Branch: master