Fix lexical enum serialization/deserialization.
We used names for enum serialization data class instances
composed from a standard string and current line number suffix
(qn_enumLexicalSerializerData_instance __LINE__).
This was working somehow in all platforms but recently I started
getting asserts in th Android app about it could not deserialize
some enum values of resource_data.json.
Investigation showed that when we call QnJson::deserialize on that fields
fusion picks up data class for deserialization from another enum.
The reason was in conflicting names of data class instances which both
were instantiated on the same line.
To solve the problem I refactored lexical enum a bit to make it use
enum name instead of current line number.
It still does not guarantee that we won't have any conflicts,
because we can create multiple enums with the same name,
but this should not happen.

<==== Mercurial data ====>
Branch: dev_2.6