Remove cotire cmake module.
Investigated what it for.
It performs two automatic optimizations: precompiled header generation and single cmopilation unit generation.
We don't use the first because we have our own precompiled headers.
The second could lead to the situation when
it's not trivial to understand where cause of compilation errors are located.
So I think we should avoid usages of such optimizations.

<==== Mercurial data ====>
Branch: cmake_experiments