VMS-4109: Item placeholder disappears on high scene zoom levels.
In the previous implementation the whole status title (icon + caption)
was emplaced in a single proxy widget wich occupied all the available space.
It led to infinite growth of that proxy widget with growth of the scene scale.
proxy widget uses grabWidget to get a pixmap, thus it tried to get
too big pixmap to render.
Now this caption cannot be bigger than it's sufficient maximum size and
we don't get such big pixmaps.

<==== Mercurial data ====>
Branch: dev_3.0.0_gui