MOBILE-402: Reimplement Switch check-state handling
The original QtQuick Controls 2 Switch Item does not provide
information if "checked" property was changed by a user interaction
or programmatically (QTBUG-57607). This maked impossible doing
things only when the user clicks a Switch. As an example it led to
a bug when Mobile Client restarts Lite Client just because when
Mobile Client lost connection for a moment, the switch state was
set to "off" (because the server has gone offline). Now it is possible
to react on user actions in the Switch.onClicked handler.
This commit can be backed-out once the mentioned bug is fixed.
However a tiny code tuning could be necessary.

<==== Mercurial data ====>
Branch: release_3.0