MOBILE-313: Fix timeline jumps after kinetic scrooling
Now Player immediately emits the signal if setPosition is called.

Previously positionChanged was emitted only after the actual frame
was played. This behavior break functioning of Binding QML item
which is used to update timeline position from player position.
"value" property of Binding updates only when player "position"
is changed. It does not update the value when Binding "when"
condition is changed. So when the timeline stops moving and setting
player position last time the Binding becomes enabled but "value"
still has an old position. Player does not report its change.
This position goes to timeline and makes it temporary jump to a location
which was passed a second ago. Then player starts playing and moves
timeline back.

Immediate emit of positionChange avoids this situation because
"value" is updated in time.

<==== Mercurial data ====>
Branch: dev_3.0.0_gui