VMS-5335: Workaround for failing updates installation when could auth is used
We noted that in multi-server system linked to the cloud
the following situation could happen:
After all update files are uploaded Client sends /api/installUpdate request
to each server. Some servers could handle this request faster than others
and go to restart. If we have cloud system and some of the "slow" servers
does not have internet access it will try to authenticate the request via
any neighbour server. But all neighbours with internet access could be
already rebooting and won't help. In this case the servers replies with 401.

In 3.1 we'll have offline authentication for such requests, but so far
it simpler to workaround the issue by allowing non-auth requests of
/api/installUpdate. To make it a bit safer Server should remove the
installed update after restart. If it doesn't do that there will be
a vulnerability which will allow somebody infinitely restart Server
by sending it requests to install the downloaded update.

<==== Mercurial data ====>
Branch: dev_3.0.0_gui