VMS-6216: Fix Lite Client camera switching behavior
Several Lite Client components use WheelSwitch to limit a number of
consequent camera switches. It wasn't working properly.
The problem was caused mainly by a bug which multiplied elements in the
same position in the layout. E.g. it was possible to have 50 items with
{0, 0} position. To avoid such bug in the future LiteClientLayoutHelper
removes all elements with the given coordinates instead of the first found.

Also a minor network optimization is done. LiteClientLayoutHelper now uses
PendingOperation to avoid frequent saveLayout transactions while the user is
frequently switching cameras (e.g. with a mouse wheel).

<==== Mercurial data ====>
Branch: vms_3.1_gui