MOBILE-622: Fix video positioning
The removed code was needed when ZoomableFlickable was able to position video
on the screen with 1/3 and 2/3 vertical paddings. Now this is done only by
VideoPositioner. One of the recent commits fixed implicitSize of
MultiVideoOutput and ZoomableFlickable tried to position it using the real
video size and expectedly failed.

This fix forces correct explicit size to VideoPositioner.

<==== Mercurial data ====>
Branch: mobile