# Replace std::round with qRound because not all compilers provide std::round
Mainly this is bacause of GCC-7.2.0 for ARM Linux build with glibc-2.13.
This version of glibc is required by Debian 7 which is our main bpi OS so far.
This old glibc does not have some mathematical functions which GCC wants, so
GCC just disables a bunch of mathematical functions which should be redefined
in std namespace.

<==== Mercurial data ====>
Branch: vms_3.2_dev