CMake: Remove qt.conf from resources of common library
For maven build nothing changed. The file just lays in another place of
sources. However CMake does not use that place. Also CMake does not override
qt.conf in resources. Instead CMake creates qt.conf in bin directory, so the
programs can still be run.

For cross-builds qt.conf contains "Prefix = ..": the same content as which
should be on a target device, thus the developer can mount the build directory
to a device and run the application. For non-cross builds qt.conf has
"Prefix = <path-to-qt>", so the application can be run from the build directory
itself.

<==== Mercurial data ====>
Branch: vms_3.1