CMake: Refactor packages versions detection
Now CMake don't store package versions in the cache, so when somebody updates
a required version of a package the new package will be used with no need to
clear CMakeCache.txt.

Note this does not apply to Qt because Qt creates a lot of its own cache
variables.

<==== Mercurial data ====>
Branch: vms_3.1