CMake: Create and use a temporary keychain for codesigning Client in MacOS
Since this commit MacOS distribution building script is able to use p12 certificates
automatically whitout a need to import them into Login keychain by hand.
To acieve this a special script creates a temporary keychain nx_build and imports
the required certificate into it. DMG building script is modified to use this keychain
when it is available.

Certificates are searched in RDEP packages "any/certificates-${customization}" in the "macosx"
directory.

<==== Mercurial data ====>
Branch: vms_3.2