CMake: Move include(utils) back to the common includes list
utils.cmake sets some variables related to CMake multi-configuration mode. The related CMake
variables are initialized by the root project() call, so utils.cmake must be included after it.

The script was moved to the beginning of the root CMakeLists.txt because of nx_option function.
This function is now moved to a separate script.

<==== Mercurial data ====>
Branch: vms_3.2_dev