VMS-8339: Disable -fdevirtualize for GCC
With this option enabled (by -O2) GCC generates broken code in some scenarios. E.g. in
QnForwardingInstrument some functions which cast a base class QGraphicsItem to a local derivative
declared in an anonymous namespace don't have ret instructions. Thus after a call to such function
the program continues execution in a random function which is actually located after. In the
best cases this just leds to a segfault.

<==== Mercurial data ====>
Branch: vms_3.2_dev