CMake: Allow skipping configuration checks using skipConfigurationChecks variable
Since this commit all configuration checks should have an ID, e.g. the check that edge1 can be
ebabled only for digitalwatchdog configuration has an ID dw_edge1. This id can be put to
skipConfigurationChecks variable to skip this check.

Also dw_edge1 check is now auto-skipped when developedBuild == ON.

This commit does introduce only an example of how to do this and does not provide any shortcuts.
This is to avoid messing with arguments escaping what would be necessary it there's a function
which accepts a condition as an input parameter.

<==== Mercurial data ====>
Branch: vms_3.2_dev