VMS-9255: metadata_sdk: Supported decoding video frames to various RGB formats; related fixes
Fixed incorrect processing of video frames in case of more than one CameraManager.
Moved classes that are not needed for plugins out of nx_sdk.

Added deep-copying of uncompressed frame during its decoding in the Motion engine (will be
optimized when the Motion engine is refactored to use AbstractVideoDecoder).

<==== Mercurial data ====>
Branch: vms