VMS-10104: Fix incompatible server configuration after update
When connecting an outdated foreign server into the current system its resource disappears when the
server restarts and then QnConnectToCurrentSystemTool cannot find the disappeared resource to
configure. Actually the resource is just re-created, so the tools should find a new resource with
the same original ID and use it.

By the way some logs are added.

<==== Mercurial data ====>
Branch: vms_3.2