VMS: In UNIX-like systems assertions send SIGTRAP signal instead of simulating SEGFAULT
This approach has a major benefit in comparison to the old one. When a program is run under
a debugger it catches the signal like a breakpoint. The program can then be continued if needed.
When the program runs without a debugger it crashes as usual and a core dump is generated.

<==== Mercurial data ====>
Branch: vms