VMS-10422: Fix update requests authentication when connecting an incompatible server
This commit brings back QnFakeMediaServerResource::setAuthenticator method removed in 04fbe19afdc9.
The authentication info is required for update requests. However the authenticator is reset when
update is finished because it causes problems when the server is being configured. The
configuration tool uses its own way to provide the credentials.

<==== Mercurial data ====>
Branch: vms_3.2_release