FT-78: Make runner address, as seen from VMs, an pre-configurable option of Hypervisor
Currently, when the address is known in advance and used to construct URLs, there is no way to
perform a port mapping. This is not a problem though, because in VirtualBox, ports are not changed
when connecing from VM to host. This, therefore, deprecates os_access.port_map.local as only
address is used from it, not port map.

<==== Mercurial data ====>
Branch: vms