Compile fix for Clang: Now it's possible to use components of structured bindings in NX_ASSERT
In Clang it's not allowed to capture components of structured binding:

struct A { int a; };
auto [a] = A();
[&a]{ std::cout << a; }() //< Compile error.

Formerly NX_ASSERT was implemented as a lambda function which captures all local variables (like
this: [&]() { ...  }). Thus the condition and message arguments were implicitly captured. This
approach led to compile errors when a component of a structured binding was used in NX_ASSERT:

struct A { int a = 1; };
auto [a] = A();
NX_ASSERT(a == 1); //< Compile error.

Now the lambda in NX_ASSERT does not capture anything, it performs assert check right in the
capture list and return the result in the body. In a simple case (when assert timers are disabled)
lambda is even not necessary, it is only used to silence warnings about unused exression result.

<==== Mercurial data ====>
Branch: vms