META-47: Fix device agent settings values after Camera Settings dialog is re-opened
When the dialog is re-opened its state is "partially" reset. In particular loaded settings values
are cleared, but current camera is preserved. This leads to default values are shown because actual
values are loaded only when user selects an engine.

Now values are loaded when either current engine or current camera is changed. Also current engine
ID is moved to the dialog state from QML file. Now the DeviceAgentSettingsAdapter monitors this
field and refreshes values when needed.

<==== Mercurial data ====>
Branch: meta