VMS-9844: Refactor cloud instance checks in the DEB package scripts
Previously the compatibility check was in the preinst script, but rollback after failed update
wasn't implemented properly. This led to broken server installation after an attempt to install a
package for another cloud instance when server is linked to a cloud.

Currently the check is moved much earlier (into prerm). This allows us to avoid any destructive
actions until we sure that the new server could be installed. To make the check possible in prerm
package naming was changed too. Previously we had only networkoptix-mediaserver package. Now we
have unique packages for all cloud groups:
    prod: networkoptix-mediaserver
    test: networkoptix-mediaserver-test
    demo: networkoptix-mediaserver-demo
    dev: networkoptix-mediaserver-dev
All packages provide "networkoptix-mediaserver" and conflict with each other. When we install a
package for another cloud instance, prerm script sees the conflict and does the check.

When a package with another cloud instance is installed, user is asked if she wants to keep or
remove the database. If the choise is "remove", then cloud instance check is skipped because
after installation the system will not be configured, and thus not linked to cloud.

Due to a dpkg feature, it can't silently replace an installed prod package with another package.
To do this extra arguments are required:
    dpkg -i -B --force-conflicts nxwitness-server-4.0.0.0-linux64-beta-test.deb
For all other cases (when you have a non-prod package installed) you can invoke dpkg as before:
    dpkg -i nxwitness-server-4.0.0.0-linux64-beta-prod.deb

<==== Mercurial data ====>
Branch: vms_4.0_gui