VMS-12526: Analytics objects overlay mechanics reimplemented
"Analytics" icon was removed from the camera widget. Instead the new context menu items were added,
which now control the anayltics mode. Objects are displayed always (by default), and this behavior
can only be switched to mode, when objects are displayed in sync with the right panel.

Setup state is persistent and bound to the layout item guid in the current system, stored locally.

<==== Mercurial data ====>
Branch: vms_4.0_gui