SDK: Migrated analytics_sdk to newly introduced nx::sdk::IRefCountable et al.; related improvements
Introduced new classes for the SDK as the binary-compatible replacements for old classes:
- nxpl::PluginInterface -> nx::sdk::IRefCountable
- nxpt::CommonRefCounter -> nx::sdk::RefCountable
- nxpl::Plugin2 -> nx::sdk::IPlugin
    - setSettings() removed - the new Plugins do not support receiving Server's roSettings
    - setPluginContainer(nxpl::PluginInterface*) -> setUtilityProvider(nx::sdk::IUtilityProvider*)
- nxpl::TimeProvider -> ITimeUtilityProvider
    - uint64_t millisSinceEpoch() -> int64_t vmsSystemTimeSinceEpochMs()

Introduced new helper class:
- nx::sdk::Interface - Helper template for defining interfaces, which implements queryInterface()

Migrated analytics_sdk to the new classes. Note that old binaries of Analytics Plugins are not
compatible with the new VMS, which is fine because prior to the upcoming VMS 4.0 Analytics Plugins
existed only as a part of the VMS code base (they were not written by third-parties).

NOTE: Video Source (Camera) SDK and Storage SDK did not migrate yet to the new classes, but after
they do, the old classes can be deleted from the repo. Such migration should be done preserving
compatibility with old plugin binaries.

<==== Mercurial data ====>
Branch: meta_4.0