VMS-13038: Display "Get Update File" button only when it makes sense
In general, if Client successfully received update information and then noted that it is invalid,
then nothing will change if a user downloads a file manually and try again. The result will be the
same. Thus the button is displayed only in case of network errors. "noError" displays a button too
just to allow user to download an update somewhere else (e.g. if she has paid traffic on Client
which is OK to check for updates but not OK to download them).

<==== Mercurial data ====>
Branch: vms_4.0_gui