Refactor how software decoding is forced for iOS devices
The conditions when we disable HW decoding tends to grow in number, thus instead of building
complex conditions we can use just a blacklist. Thus when we have a request, we'll just add a new
item into it. Also such list is much simpler to read when it's needed to understand which deviced
are black-listed.

<==== Mercurial data ====>
Branch: mobile