Refactor Audio Recording settings
Now it does not request all available devices from Qt, instead the list is cached and re-used. This
speeds up how settings are loaded a bit.

Also when a user selects "None" it no longer stores tr("None") in settings. The specific string
"<none>" is stored for that. Settings migration is not implemented (the decision to not do it was
discussed). If in an older version "None" was selected, then after upgrade the default option will
be chosen.

Another improvement is an ability to reset settings to "auto". Now "Auto" is included in the list
of options and could be selected.

<==== Mercurial data ====>
Branch: vms_4.0_gui