SDK: Remove dangerous Ptr::dynamicCast(); add null check to queryInterfacePtr()
The removed Ptr::dynamicCast() method was dangerous because it used dynamic_cast<>() for
ref-countable objects which is unsafe: dynamic_cast<>() can be ABI-dependent and should not be used
at the Server side for objects created in a plugin and vice versa. Instead of the removed method,
it is recommended to use queryInterfacePtr().

Previously, queryInterfacePtr() would crash if a null pointer was supplied as its argument. Now, it
just returns null smart pointer (Ptr<>) in this case (as it does when the object does not support
the requested interface).

<==== Mercurial data ====>
Branch: meta_4.0