VMS-13764: Update handling of peers in 'unknown' state during system update
All state of update participants will change to 'unknown' when user clicks 'Download' button. This state will
be changed when client gets /ec2/updateStatus. This state is displayed as 'waiting for server data... ' status
in component table.

There will be an error when server stays in this 'unknown' state for too long.

<==== Mercurial data ====>
Branch: vms_4.0_gui