Refactor peer managers / peer selection in p2p Downloader
Previously in downloader and its peer managers was a concept of direct internet download in
contrast to peer download. This worked not well with a concept of peer rank which was applied to
other peers only, not internet. Now "peer" in the downloader is a pair of peer manager and peer ID,
and direct internet downloads are done via separate peer managers (and thus separate peers). Thus
now internet downloads are also ranked.

Also some things were simplified:
- peer selectors (abstract and derived) were joined into one small class,
- peer manager factories were removed, now downloader just accepts a list of certain peer manager
instances and passes them to all workers.

<==== Mercurial data ====>
Branch: vms_4.0_gui