VMS-13894: Introduce error code for nx::update::Status
Previously detailed information about errors was stored in `message` and was displayed directly
from it in the Client. The approach is invalid because does not take translations into account.
Now error information is stored in a separate errorCode enum field, and Client could display a
translated message based on the enum value.

<==== Mercurial data ====>
Branch: vms_4.0_gui