VMS-13475: Fix installation of Server dependencies for linux_arm32
cifs-utils is not installed by default in Ubuntu image for Banana Pi and installation of this
dependency was broken. install.sh script was fixed for this.

linux_arm32 can be installed on debian-like and ubuntu-like systems both. The logic of distinction
between this two cases was added to install.sh. Corresponding artifact cifs-utils was updated.

<==== Mercurial data ====>
Branch: vms_4.0_server