VMS-13822: Fix memory leak in Server plugins of UtilityProvider aka PluginContainer
The leak is caused by UtilityProvider (former PluginContainer) object that survives Server
shutdown. This reference-countable object is owned by PluginManager, and is passed to each plugin.
The plugins which are part of our VMS increment UtilityProvider’s reference counter, but never
decrement it (though should). In 3.2 this did not cause a leak because the UtilityProvider object,
though reference-countable, was owned in a PluginManager field by value (which was obviously a bad
practice as well). Thus, the superposition of the two contract violations in 3.2 has led to no
visible issue.

Decided to fix all VMS plugins in 4.0 to decrement UtilityProvider’s reference counter.

ATTENTION: If an old, non-decrementing, plugin (e.g. written by a third-party) is used with the 4.0
Server, the leak will be there, but it is not dangerous (the object’s lifetime is the Server’s
lifetime anyway).

<==== Mercurial data ====>
Branch: vms_4.0_server