VMS-12530: Fix segfault when combo box menu is opened from WebKit in Linux
The segfault is caused by a dynamic cast from the WebKit internal QtWebComboBox which is built with
RTTI disabled. However GCC implementation of dynamic cast uses RTTI, so the program crashes when
trying to dereference a null pointer (RTTI data).

The simplest solution here is to use qobject_cast which does not depend on RTTI. However we
probably ought to re-build Qt WebKit with RTTI enabled for Qt part.

<==== Mercurial data ====>
Branch: vms_4.0_gui