VMS-14611: Avoid banning peers in downloader which had several failed outdated requests
There's a case when due to a temporary network outage several requests in downloader hang for
a long time and then fail one after another. This makes the peer banned even if it successfully
processed a lot of new requests when network issues are gone. Now worker tracks the timestamp of
the latest successful request, and don't decrease peer rank if notes that a failed request was
issued before any of successful requests.

<==== Mercurial data ====>
Branch: vms_4.0_gui