VMS-9816: Fix deduction of the server primary interface
In the very beginning we setup default url as the address we used to connect (without concerning
forbidden addresses). After that moduleFinder should guest the best interface for the server and
update the current setup (asynchronously).In some rare this could get to races, when module finder
finds default address before server is added to the resources pool.

<==== Mercurial data ====>
Branch: vms_4.0_gui