Improved analytics objects search consistency
Search by text now is implemented using the same syntax as in SQLite fts4, so only full text match
and prefix search are supported. When user enters something in the search field, it is interpreted
as a prefix search by each word separately. Asteriks is appended to the API search call on the
client side.

Only attribute values are now stored in the FTS database and therefore are available for search.

Fixes META-393. Should also fix META-392 and META-385.

<==== Mercurial data ====>
Branch: vms_4.0_gui