VMS-17035: Fix analytics settings loading
Added a lot of comments and log strings to the analytics settings classes (mostly to the manager).
Settings listening policy is now set in constructor and cannot be changed. Only compatible engines
are now listened by the camera settings dialog. Repeating settings refresh requests are not sent
if not actually needed.

Branch: vms_4.1