VMS-17073: ROI GUI behavior update
The name is filled but the shape is not created, it's possible to click
Apply and save config.

If figure is empty, "Display on Camera" checkbox is disabled.

After creating a figure and closing the window, automatically enable
checkbox and put it in the "Clicked" status.

If you delete a figure, the checkbox will be disabled and the name will
be deleted.

Branch: vms_4.1