VMS-17953: Rewrite Mediaserver config helper from python to bash
This allowed us to remove python dependency from Mediaserver which was used only for this script.
The script was compatible to both Python 2 ann Python 3 and invoked Python as `python`. But
Ubuntu 20 does not provide it, so probably it's time to remove python dependency at all. We could
just switch to python3, but even though all supported Linux distros have python3 package, we cannot
be sure that all existing systems have it installed. So it's better to do simple tasks with simple
tools and don't carry extra dependencies.

Branch: vms_4.1