VMS-18318: Clear extracted update files only when new download is started
Previously server used to remove update files immediately ater its new version is started. The
problem is that sometimes server does it so fast, that it does it before update script is finished.
Furthermore it does it before dpkg finished working. This produces unexpected errors in the update
script, because the files disappeared.

Now update files will be removed only when server starts downloading a new version. Thus the update
script should finish with no errors.