VMS-11738: nx_kit: Support std::string with '\0' in toString() and ASSERT_STREQ()
Support std::string with `\0` as an in-string item in toString() and ASSERT_STREQ().

In toString() for strings (performs C-style-escaping), do not produce empty quotes after the
generated `\u` and `\U` escape sequences. Note that the empty quotes are still needed after `\x` to
guarantee the termination of the hex sequence, but `\u` and `\U` are limited to 4 or 8 hex digits
respectively, and toString() generates them with enough leading zeroes.