VMS-18697: Fix client crash when opening local file using file manager
`ClientStartupHandler` logic started to work before we have initialized the startup parameters. So
there was a race when a password-protected file was opened. Connecting was queued and triggered
while password dialog started it's event loop.

Now logic is initialized externally, when we definitely know if we are connecting to the system or
opening a local file, etc.