VMS-19166: Fix Client crash when last layout is deleted
The crash happenes because when action is triggered directly it derives action parameters from an
action that was triggered higher in stack. In this case action derives a list with a deleted
layout (last closed layout). Access to a deleted instance obviously leads to a segfault.

To avoid this issue action should be triggered via `menu()->trigger()` method which sets the
parameters.