VMS-19388: Make it possible to run 4.1 or newer Client from 4.0 or older Applauncher
Recently we've added libstdc++ conditional loading into the `client` script. This made it possible
to run Client successfully in any of supported Ubuntu versions using this script. Applauncher was
configured to use this script instead of calling the Client binary directly. However old
Applauncher still tries to load the binary (client-bin). The solution is to make `client-bin` a
symlink to `client` script for backwards compatibility and rename the binary to something else.
The new binary name contains brand information and it is now `nxwitness_client` for the default
customization.