VMS-18605: Validate update packages before extraction or installation
After this commit Client and Server will check a digital signature of every update package that is
going to be installed. The signatures must be provided in the update information (packages.json).

For simplified development and testing, it is possible to disable verification by a corresopnding
option in the INI-configs.