VMS-19673: Remove camera advanced params manifest from DB
`CameraAdvancedParams` is the manifest for the advanced camera settings dialog. It is quite heavy
and was unnecessary made as the "cameraAdvancedParams" resource property. On systems with many
cameras this property leads to JSON serialization fail of `ec2/getCamerasEx` response due to QJson
size limit. This commit removes the property from DB and adds the `api/getCameraParamManifest` API
endpoint to request the manifest by the dialog on demand.