VMS-20016, VMS-20040: Fix font weight in some places
For some reason Qt Quick sometimes renders some text elements with wrong font weight if the weight
is not set explicitly (text becomes too light). It looks like this is Qt 5.11 issue which is fixed
in Qt 5.14. So it's simpler to just manually set the correct weight now instead of digging in Qt
sources to find the reason.