VMS-20242: fix resource data.json hikvision
Many records for Hikvision are duplicated for lowercase and uppercase spelling of `Hikvision` vendor. Uppercase keys should be eliminated.
Some records have "hangzhou hikvision digital technology co., ltd" vendor name as a key, this is an error as vendor name is shortened with 'm_shortVendorByName' map that has a pair "hangzhou hikvision digital technology co., ltd" -> "hikvision". So the key for such records should be changed to "Hikvision".