Analytic SDK camera timestamp translation (master)
Add `cameraClockTimestamp` flag to `nx::sdk::analytics::IDataPacket`s to indicate that timestamps of packets going from server to plugin are from camera's own clock and the server should translate them to VMS system clock.

This has been split off from !2293 as it turned out to be unnecessary there, but @rvasilenko considers it to be useful on its own.