VMS-21752: fix proxy authorization in HTTP client
* Http client authentication cache (`http::AuthInfoCache`) is re-written and now works properly. Also, it was simplified greatly
* added missing tests that check authentication on a proxy and reusing server's nonce to pass Digest authentication in one round trip if the server allows it

@muskov @rvasilenko

After I properly implemented server nonce reuse to enable one-RTT Digest authentication, the following server tests broke:

- EventOrderTest/eventsFromMultipleServersAreOrderedByTimestamp
- MetricsServersApi.threeServers

It turned out that the CONNECT proxy-reverse handler does not send the WWW-Authenticate header in the "401 Unauthorized" response if the request contained Authorization. Which is strange and different from all other handlers.
I made these tests working by adding httpClient->setDisablePrecalculatedAuthorization (true); in reverse_connection_manager.cpp.
But, of course, it's worth fixing it in the server.

- AuthenticationTest/sessionKey
- AuthenticationTest/lockoutTest

The problem is that after enabling pre-authorization in the HTTP client (which, as it turned out, almost never worked) the Authorization
header is delivered in the first request and is immediately counted as an unsuccessful authorization attempt.
So far I've solved it by adding httpClient->setDisablePrecalculatedAuthorization (true); in createHttpClient (test_api_requests.cpp).
This should be supported in the server so that it does not consider an attempt to reuse a nonce as an invalid password entry.