VMS-31315: Fix restoring credentials after reboot on Ubuntu
If autologin on Ubuntu is configured, keyring is not initialized automatically when logging in. Keyring does not work until the user enters the master password manually. But client's attempt to read the password stored in the keyring does not initiate the input of the master password by the user. Only an attempt to store the password initiates entering the master password. Even with empty both the key and the value.