VMS-31473: Fix available analytics event types selection in the Right Panel
Taxonomy in not yet ready for events (and won't be in 5.0)

Besides, we take into consideration not only active event taxonomy, but also existing event rules.

So I reverted to old signal notifications and kept multi-threaded optimization.

I renamed some methods and removed `Q_INVOKABLE` - is it OK? (It's not OK if it was intended for Squish testing, for example)

Closes VMS-31473