VMS-34510, VMS-34512: Support Cross-system cameras in the Resource Tree
Additional node is added to resource tree under cloud systems. Ite can work as a preloader or
display user interaction request. When clicked, connection is re-establised with allowed user
interaction.

Cameras on cross-system layouts are also displayed as preloaders while connection is not
established. If such layout is opened on the workbench, cameras are automatically added when
they are available.

Fixes VMS-34510, fixes VMS-34512.