VMS-34779: Support new `flavor` field in system updates
This new field is intended to be customized by vendors to create their own distributions of the VMS
Server. For system updates the `flavor` works in the same way as the `platform` field: update
package is considered incompatible if its flavor does not match the flavor of the current system.

Flavor can be customized by placing the property `currentOsFlavor` into the Server config file
(mediaserver.conf).