VMS-35041: ARM-specific behavior: Introduce Server setting ignoreUsb0NetworkInterfaceIfOthersExist
In nx_network, when enumerating network interfaces in getAllIPv4Interfaces(), there was a clause on
ARM32 devices to ignore the "usb0" interface if it was not the only one. Now, this clause is
activated by the new Server setting which is turned off by default.

NOTE: The rationale for the original solution is unclear, probably it was needed for some obsolete
ARM device.

(cherry picked from commit ad2a5ada343c252dd4f24c026f364da0afc6c6a5)