VMS-35041: ARM-specific behavior: Introduce Server setting sleepIfEmptySocket
On Linux, on ARM32 devices there was a special behavior when reading an RTSP stream(in QnRtspClient
defined in nx_vms_common/src/nx/streaming/): when reading from a socket, sleep for a while in case
no data is available. Now, this clause is activated by the new Server setting which is turned off
by default. Since there is no Client on ARM32, the Client behavior did not change.

NOTE: The rationale for the original solution is unclear, probably it was needed to lower the CPU
usage on some obsolete ARM device.