VMS-34985: ARM-specific behavior: Introduce Server setting maxTcpRequestSize
Previously this value was hard-coded in QnTCPConnectionProcessor
(open/vms/libs/nx_vms_common/src/network/tcp_connection_processor.cpp), being 512 MB for Windows
and the linux_x64 platform, and 16 MB (16777216) for the linux_arm32 and edge1 platforms.

The Server behavior did not change, because the value 16777216 was added to mediaserver.conf for
the linux_arm32 and edge1 platforms.