VMS-34985: ARM-specific behavior: Introduce Server setting resolveLocalhost
The new setting instructs nx_network to manually resolve the localhost address, as was previously
hard-coded for the `edge1` platform. This is a workaround for operating systems where /etc/hosts
does not work, thus localhost is inaccessible, like certain Busybox-based systems.

The Server behavior did not change because this option with the value "true" was added to
mediaserver.conf for the edge1 platform.