VMS-34985: ARM-specific behavior: Introduce Server setting threadStackSize
The new setting allows to reduce the thread stack size (from the default 8 MB on Linux) to save
memory on low-profile devices like ARM. Previously the value of 128 KB has been hard-coded for the
`edge1` platform.

The Server behavior did not change, because this option with the value 131072 was added to
mediaserver.conf for the edge1 platform.