VMS-34985: ARM-specific behavior: Use Server setting disableTranscoding as the only mean
Previously transcoding was additionally disabled for ARM32 devices. Also, this Server setting did
not lead to the Server flag SF_SupportsTranscoding being turned off. Now the logic is simple -
transcoding is enabled if and only if disableTranscoding is false, and SF_SupportsTranscoding
directly reflects that.

Also, removed the obsolete code which allowed a Client to additionally disable transcoding on its
side via calling VideoDecoderRegistry::setTranscodingEnabled() - it has been used only in Lite
Client (which is already removed).

The Server behavior did not change because this option with the value "true" was added to
mediaserver.conf for the edge1 and linux_arm32 platforms.