VMS-34985: ARM-specific behavior: Introduce Server setting isEdgeServer
Previously the Server flag SF_Edge has been set for and only for the edge1 platform, via the
isEdgeServer CMake variable. Now this CMake variable is removed, and the Server flag is set based
on the new setting.

The Server behavior did not change, because this option with the value "true" was added to
mediaserver.conf for the edge1 platform.